project/requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile project/requirements/base.in
#
amqp==5.2.0
    # via kombu
asgiref==3.7.2
    # via django
billiard==4.2.0
    # via celery
celery[redis]==5.3.6
    # via
    #   -r project/requirements/base.in
    #   pytest-celery
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
django==5.0.2
    # via
    #   -r project/requirements/base.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r project/requirements/base.in
gunicorn==21.2.0
    # via -r project/requirements/base.in
idna==3.6
    # via requests
iniconfig==2.0.0
    # via pytest
kombu==5.3.5
    # via celery
mock==5.1.0
    # via -r project/requirements/base.in
packaging==23.2
    # via
    #   gunicorn
    #   pytest
pluggy==1.4.0
    # via pytest
prompt-toolkit==3.0.43
    # via click-repl
pytest==8.0.2
    # via
    #   -r project/requirements/base.in
    #   pytest-django
pytest-celery==0.0.0
    # via -r project/requirements/base.in
pytest-django==4.8.0
    # via -r project/requirements/base.in
python-dateutil==2.8.2
    # via celery
python-dotenv==1.0.1
    # via -r project/requirements/base.in
pytz==2024.1
    # via djangorestframework
redis==5.0.1
    # via celery
requests==2.31.0
    # via -r project/requirements/base.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tzdata==2024.1
    # via celery
urllib3==2.2.1
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
